requirements.txt
asgiref==3.4.1
autopep8==1.6.0
certifi==2021.10.8
charset-normalizer==2.0.12
dj-database-url==0.5.0
Django==3.2.11
django-crispy-forms==1.13.0
django-filter==21.1
flake8==4.0.1
gunicorn==20.1.0
idna==3.3
importlib-metadata==4.2.0
isort==5.10.1
mccabe==0.6.1
Naked==0.1.31
pycodestyle==2.8.0
pyflakes==2.4.0
python-decouple==3.5
pytz==2021.3
PyYAML==6.0
requests==2.27.1
sqlparse==0.4.2
status==0.2.5
toml==0.10.2
typing_extensions==4.1.1
Unipath==1.1
urllib3==1.26.9
whitenoise==5.3.0
zipp==3.6.0
